Enhancement: use commit comparsion to compare commits
