+++ core/commands/log.py
@@ -233,98 +233,3 @@ class GsLogBranchCommand(WindowCommand, GitCommand):
             return
         self._selected_branch = self.all_branches[index]
         self.window.run_command("gs_log", {"branch": self._selected_branch})
-
-
-class GsCompareAgainstCommand(WindowCommand, GitCommand):
-    def run(self, target_commit=None, file_path=None):
-        self._file_path = file_path
-        self._target_commit = target_commit
-        sublime.set_timeout_async(self.run_async)
-
-    def run_async(self):
-        options_array = [
-            "Any Reference",
-            "Branch"
-        ]
-
-        self.window.show_quick_panel(
-            options_array,
-            self.on_select_against,
-            flags=sublime.MONOSPACE_FONT,
-            selected_index=self.quick_panel_compare_against_idx
-        )
-
-    def on_select_against(self, index):
-        if index < 0:
-            return
-
-        self.quick_panel_compare_against_idx = index
-
-        if index == 0:
-            self.window.run_command("gs_compare_against_reference", {
-                "target_commit": self._target_commit,
-                "file_path": self._file_path,
-                "from_panel": True
-            })
-
-        if index == 1:
-            self.window.run_command("gs_compare_against_branch", {
-                "target_commit": self._target_commit,
-                "file_path": self._file_path,
-                "from_panel": True
-            })
-
-
-class GsCompareAgainstReferenceCommand(WindowCommand, GitCommand):
-    def run(self, target_commit=None, file_path=None, from_panel=False):
-        self._file_path = file_path
-        self._target_commit = target_commit
-        self.from_panel = from_panel
-        sublime.set_timeout_async(self.run_async)
-
-    def run_async(self):
-        self.window.show_input_panel("Ref:", "", self.show_diff, None, self.on_cancel)
-
-    def show_diff(self, ref):
-        self.window.run_command("gs_diff", {
-            "file_path": self._file_path,
-            "current_file": bool(self._file_path),
-            "base_commit": ref,
-            "target_commit": self._target_commit,
-            "disable_stage": True
-        })
-
-    def on_cancel(self):
-        if self.from_panel:
-            self.window.run_command("gs_compare_against", {
-                "target_commit": self._target_commit,
-                "file_path": self._file_path
-            })
-
-
-class GsCompareAgainstBranchCommand(GsLogBranchCommand):
-    """
-    Compare a given commit against a selected branch or selected ref
-    """
-    def run(self, target_commit=None, file_path=None, from_panel=False):
-        self._file_path = file_path
-        self._target_commit = target_commit
-        self.from_panel = from_panel
-        sublime.set_timeout_async(self.run_async)
-
-    def on_branch_selection(self, index):
-        if index < 0:
-            if self.from_panel:
-                self.window.run_command("gs_compare_against", {
-                    "target_commit": self._target_commit,
-                    "file_path": self._file_path
-                })
-            return
-        self._selected_branch = self.all_branches[index]
-        self.window.run_command("gs_diff", {
-            "file_path": self._file_path,
-            "current_file": bool(self._file_path),
-            "base_commit": self._selected_branch,
-            "target_commit": self._target_commit,
-            "disable_stage": True
-        })
